lab2/homework/requirements.txt
fastapi==0.110.0
httpx==0.27.0
python-dotenv==1.0.1
Requests==2.31.0
